requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.1.14
django-cors-headers==4.9.0
django-leaflet==0.32.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
GDAL==3.10.2
idna==3.11
pillow==12.0.0
psycopg==3.2.12
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
